Added optional edit links to multi selects
Links have been added optionally to the multi selects which basically
allow users to quickly access page sections.  This can be extended going
forward to use a templated link that uses the "key => value" of select
options to create a dynamic list of links or any content based on the
select.  Once implemented properly this could be the basis for some
pretty decent image previews too.