Major ListTaxonomy Update
This function now uses {codes} instead of requiring the user to use the
clunky sprintf variables %1$s etc...

A wrap_template with <li> as a default has also been added to alleviate
issues with the wyziwyg wanting to add a <ul> when switching from visual
to html mode.

The terms output has been modified to use the new string tools function
to output the comma delimited list of terms.

The taxonomy is output using the OutputPostProperties function added to
the date tools file.  Down the road the function could be modified to
include the base post fields as well as dynamic properties included from
the database.

Lastly a "AddSidebar" shortcode has been added which uses a janky output
buffer.