Added framework for customizer_data
Similar to taxonomy_data and admin_data - customizer_data allows us to
define a hardcoded set of variables and defaults for a theme to use / be
implemented by a child class.  At a later date these changes could be
rolled into the theme_core class and implemented through the child
theme's core but right now the inheritance isn't setup to allow for this
to work seamlessly.  These changes have been added into wp-tools.php for
now.